+++ apps/indexer/lib/indexer/fetcher/multichain_search_db/main_export_queue.ex
@@ -19,7 +19,7 @@ defmodule Indexer.Fetcher.MultichainSearchDb.MainExportQueue do
 
   @default_max_batch_size 1000
   @default_max_concurrency 10
-  @failed_to_re_export_data_error "Batch export retry to the Multichain Search DB failed"
+  @failed_to_re_export_data_error "Batch main export retry to the Multichain Search DB failed"
 
   @doc false
   def child_spec([init_options, gen_server_options]) do